+++ monew/src/main/java/com/codeit/team2/monew/module/domain/interest/controller/InterestController.java
@@ -1,4 +1,4 @@
-package com.codeit.team2.monew.module.controller.interest;
+package com.codeit.team2.monew.module.domain.interest.controller;
 
 import com.codeit.team2.monew.module.domain.interest.dto.InterestDto;
 import com.codeit.team2.monew.module.service.InterestService;
