+++ monew/src/main/java/com/codeit/team2/monew/module/domain/interest/controller/InterestRegisterRequest.java
@@ -1,4 +1,4 @@
-package com.codeit.team2.monew.module.controller.interest;
+package com.codeit.team2.monew.module.domain.interest.controller;
 
 import java.util.List;
 import lombok.Builder;
