+++ monew/src/main/java/com/codeit/team2/monew/module/domain/interest/dto/InterestDto.java
@@ -1,6 +1,5 @@
 package com.codeit.team2.monew.module.domain.interest.dto;
 
-import com.codeit.team2.monew.module.domain.interest.entity.Keyword;
 import java.util.List;
 import java.util.UUID;
