+++ monew/src/main/java/com/codeit/team2/monew/module/domain/interest/mapper/InterestMapper.java
@@ -1,8 +1,7 @@
-package com.codeit.team2.monew.module.mapper.interest;
+package com.codeit.team2.monew.module.domain.interest.mapper;
 
 import com.codeit.team2.monew.module.domain.interest.dto.InterestDto;
 import com.codeit.team2.monew.module.domain.interest.entity.Interest;
-import com.codeit.team2.monew.module.domain.interest.entity.Keyword;
 import java.util.List;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
