+++ monew/src/main/java/com/codeit/team2/monew/module/domain/interest/repository/InterestRepository.java
@@ -1,4 +1,4 @@
-package com.codeit.team2.monew.module.repository.interest;
+package com.codeit.team2.monew.module.domain.interest.repository;
 
 import com.codeit.team2.monew.module.domain.interest.entity.Interest;
 import java.util.UUID;
