+++ monew/src/main/java/com/codeit/team2/monew/module/domain/interest/repository/KeywordRepository.java
@@ -1,4 +1,4 @@
-package com.codeit.team2.monew.module.repository.interest;
+package com.codeit.team2.monew.module.domain.interest.repository;
 
 import com.codeit.team2.monew.module.domain.interest.entity.Keyword;
 import java.util.Optional;
